Move struct/enum in parser::entity:: to ast::entity::
