+++ espr/src/ast/mod.rs
@@ -1,5 +1,6 @@
 //! Abstract Syntax Tree (AST) of EXPRESS Language
 
 pub mod algorithm;
+pub mod entity;
 pub mod expression;
 pub mod types;
